Changed size and capacity type from int to size_t in Dynamic Array implementation

+++ Stack/Array/stack.c
@@ -7,7 +7,7 @@
 
 struct Stack {
     int *v;
-    int size, capacity;
+    size_t size, capacity;
 };
 
 Stack* createStack() {
